Reviewde Version
Hi,
One of the tutors mentioned that using a pre-existing implementation of a function or algorithm may not earn you the full points for this assignment. As a result, I've added a second solution to calculate entropy.

Additionally, I’ve updated the color scheme for Part A of Assignment 4 to ensure colors correspond correctly with the plots.